separate DataStore from index
